WIP abstract out core DB ops to crate
