+++ crates/Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["catchup_worker"]
+members = ["catchup_worker", "hn_core"]
 resolver = "2"
