+++ crates/catchup_worker/Cargo.toml
@@ -29,6 +29,7 @@ flume = "0.12.0"
 futures = "0.3.31"
 futures-util = "0.3.31"
 governor = "0.10.4"
+hn_core = { version = "0.1.0", path = "../hn_core" }
 log = "0.4.29"
 nonzero_ext = "0.3.0"
 prometheus-client = "0.24.0"
