+++ crates/hn_core/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "hn_core"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+diesel = { version = "2.3.6", features = ["postgres"] }
+diesel-async = { version = "0.7.4", features = ["postgres", "deadpool"] }
+serde = { version = "1.0.228", features = ["derive"] }
+serde_json = "1.0.149"
